backend/app/ml_core/requirements.txt
langchain==0.3.0
langchain-community
langchain-openai
langchain-google-genai
langchain-postgres
langchain-huggingface
psycopg[binary]
pgvector
pypdf
sentence-transformers
python-dotenv
# ML Analytics - Requirements
# Python 3.10 or 3.11 recommended

# ============================================================================
# CORE FRAMEWORK
# ============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# ============================================================================
# DATA PROCESSING & ML
# ============================================================================
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0

# ============================================================================
# EXPLAINABILITY
# ============================================================================
shap==0.44.1
lime==0.2.0.1

# ============================================================================
# VISUALIZATION (Backend)
# ============================================================================
matplotlib==3.8.2
seaborn==0.13.1

# ============================================================================
# HTTP REQUESTS
# ============================================================================
httpx==0.26.0
requests==2.31.0

# ============================================================================
# DATABASE (Optional - if using PostgreSQL)
# ============================================================================
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.25

# ============================================================================
# AUTHENTICATION (Optional)
# ============================================================================
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4

# ============================================================================
# LANGCHAIN (Optional - if integrating LangChain)
# ============================================================================
# langchain==0.1.0
# langchain-openai==0.0.2
# langchain-google-genai==0.0.6
# openai==1.6.1
# google-generativeai==0.3.2

# ============================================================================
# VECTOR DATABASE (Optional - if using PGVector)
# ============================================================================
# pgvector==0.2.4
# sentence-transformers==2.2.2

# ============================================================================
# UTILITIES
# ============================================================================
python-dotenv==1.0.0
pyyaml==6.0.1

# ============================================================================
# TESTING
# ============================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# ============================================================================
# DEVELOPMENT
# ============================================================================
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# ============================================================================
# JUPYTER (for notebooks)
# ============================================================================
jupyter==1.0.0
ipykernel==6.29.0
notebook==7.0.6

# ============================================================================
# ADDITIONAL UTILITIES
# ============================================================================
tqdm==4.66.1
colorama==0.4.6
